requirements.txt
accelerate==1.13.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
alembic==1.18.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
arrow==1.4.0
asgiref==3.11.1
attrs==26.1.0
audioread==3.1.0
av==17.0.0
binaryornot==0.6.0
blake3==1.0.8
certifi==2026.2.25
cffi==2.0.0
chardet==7.4.3
charset-normalizer==3.4.7
click==8.3.2
clip-interrogator==0.6.0
color-matcher==0.6.0
comfy-aimdo==0.2.12
comfy-cli==1.7.2
comfy-kitchen==0.2.8
comfyui-embedded-docs==0.4.3
comfyui-manager==4.2b1
comfyui-workflow-templates-core==0.3.209
comfyui-workflow-templates-media-api==0.3.69
comfyui-workflow-templates-media-image==0.3.127
comfyui-workflow-templates-media-other==0.3.176
comfyui-workflow-templates-media-video==0.3.78
comfyui_frontend_package==1.42.11
comfyui_workflow_templates==0.9.57
contourpy==1.3.3
cookiecutter==2.7.1
cryptography==46.0.7
cuda-bindings==13.0.3
cuda-pathfinder==1.2.2
cuda-toolkit==13.0.2
cycler==0.12.1
ddt==1.7.2
decorator==5.2.1
diffusers==0.37.1
docutils==0.22.4
einops==0.8.2
filelock==3.25.2
flatbuffers==25.12.19
fonttools==4.62.1
frozenlist==1.8.0
fsspec==2026.2.0
ftfy==6.3.1
gguf==0.18.0
gguf-node==0.2.8
gitdb==4.0.12
GitPython==3.1.46
glfw==2.10.0
greenlet==3.4.0
h11==0.16.0
hf==1.11.0
hf-xet==1.4.3
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.11.0
idna==3.11
ImageIO==2.37.3
imageio-ffmpeg==0.6.0
importlib_metadata==9.0.0
Jinja2==3.1.6
joblib==1.5.3
json-logic==0.7.0a0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kiwisolver==1.5.0
kornia==0.8.2
kornia_rs==0.1.10
lark==1.3.1
lazy-loader==0.5
librosa==0.11.0
llvmlite==0.47.0
Mako==1.3.11
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
mixpanel==5.1.0
mpmath==1.3.0
msgpack==1.1.2
multidict==6.7.1
networkx==3.6.1
numba==0.65.0
numpy==2.4.3
nvidia-cublas==13.1.0.3
nvidia-cuda-cupti==13.0.85
nvidia-cuda-nvrtc==13.0.88
nvidia-cuda-runtime==13.0.96
nvidia-cudnn-cu13==9.19.0.56
nvidia-cufft==12.0.0.61
nvidia-cufile==1.15.1.6
nvidia-curand==10.4.0.35
nvidia-cusolver==12.0.4.66
nvidia-cusparse==12.6.3.3
nvidia-cusparselt-cu13==0.8.0
nvidia-nccl-cu13==2.28.9
nvidia-nvjitlink==13.0.88
nvidia-nvshmem-cu13==3.4.5
nvidia-nvtx==13.0.85
onnxruntime==1.24.4
onnxruntime-gpu==1.24.4
open_clip_torch==3.3.0
opencv-contrib-python==4.13.0.92
opencv-python==4.13.0.92
opencv-python-headless==4.13.0.92
packaging==26.1
pathspec==1.0.4
pedalboard==0.9.22
peft==0.19.1
pillow==12.1.1
platformdirs==4.9.6
polars==1.40.0
polars-runtime-32==1.40.0
pooch==1.9.0
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==7.34.1
psutil==7.2.2
pycparser==3.0
pydantic==2.13.2
pydantic-settings==2.13.1
pydantic_core==2.46.2
PyGithub==2.9.1
Pygments==2.20.0
PyJWT==2.12.1
PyMatting==1.1.15
PyNaCl==1.6.2
PyOpenGL==3.1.10
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-slugify==8.0.4
PyYAML==6.0.3
questionary==2.1.1
referencing==0.37.0
regex==2026.4.4
rembg==2.0.75
requests==2.33.1
rich==15.0.0
rpds-py==0.30.0
ruff==0.15.11
safetensors==0.7.0
scikit-image==0.26.0
scikit-learn==1.8.0
scipy==1.17.1
semver==3.0.4
sentencepiece==0.2.1
setuptools==70.2.0
shellingham==1.5.4
simpleeval==1.0.7
six==1.17.0
smmap==5.0.3
soundfile==0.13.1
soxr==1.0.0
spandrel==0.4.2
SQLAlchemy==2.0.49
sympy==1.14.0
text-unidecode==1.3
threadpoolctl==3.6.0
tifffile==2026.4.11
timm==1.0.26
tokenizers==0.22.2
toml==0.10.2
tomlkit==0.14.0
torch==2.11.0+cu130
torchaudio==2.11.0+cu130
torchsde==0.2.6
torchvision==0.26.0+cu130
tqdm==4.67.3
trampoline==0.1.2
transformers==5.5.4
triton==3.6.0
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.1
ultralytics==8.4.40
ultralytics-thop==2.0.18
urllib3==2.6.3
uv==0.11.7
wcwidth==0.6.0
websocket-client==1.9.0
wordcloud==1.9.6
yarl==1.23.0
zipp==3.23.1
